fix: finalize path conflict update cleanup
Split oversized update/path-conflict modules into focused bricks, surface archive traversal and cache invalidation errors, and keep the path-conflict regression coverage deterministic.
