+++ crates/amplihack-cli/src/update/mod.rs
@@ -1,13 +1,15 @@
+mod archive;
 mod check;
+mod checksum;
 mod install;
 mod network;
 mod post_install;
 
+pub(crate) use archive::extract_archive;
 pub use check::{
     StartupUpdateOutcome, maybe_print_update_notice_from_args, run_update,
     should_skip_update_check_for_subcommand,
 };
-pub(crate) use install::extract_archive;
 pub(crate) use network::{fetch_branch_head_sha, http_get_with_retry, validate_download_url};
 
 use anyhow::{Result, anyhow, bail};
