+++ crates/amplihack-cli/src/update/tests/install_target.rs
@@ -2,10 +2,9 @@ use super::super::install::{
     BinaryInstallPlan, InstallArchiveLayout, plan_downloaded_binary_install,
 };
 use crate::path_conflicts::{
-    BinaryProbe, InstallTargetDecision, PathAnalysisInput, TargetDecisionInput,
-    analyze_path_conflicts,
+    InstallTargetDecision, PathAnalysisInput, TargetDecisionInput, analyze_path_conflicts,
+    probe_candidates_without_exec,
 };
-use std::collections::BTreeMap;
 use std::fs;
 use std::path::{Path, PathBuf};
 
@@ -31,8 +30,8 @@ fn downloaded_update_plan_uses_preferred_user_bin_when_system_binary_shadows_cur
     let archive = temp.path().join("archive");
 
     let current_exe = write_executable(&usr_local_bin, "amplihack");
-    let user_amplihack = write_executable(&user_bin, "amplihack");
-    let user_hooks = write_executable(&user_bin, "amplihack-hooks");
+    write_executable(&user_bin, "amplihack");
+    write_executable(&user_bin, "amplihack-hooks");
     let new_amplihack = write_executable(&archive, "amplihack");
     let new_hooks = write_executable(&archive, "amplihack-hooks");
 
@@ -43,32 +42,10 @@ fn downloaded_update_plan_uses_preferred_user_bin_when_system_binary_shadows_cur
         binary_names: vec!["amplihack".into(), "amplihack-hooks".into()],
     })
     .unwrap();
-    let mut probes = BTreeMap::new();
-    probes.insert(
-        current_exe,
-        BinaryProbe {
-            version: Some("0.9.60".into()),
-            writable: false,
-        },
-    );
-    probes.insert(
-        user_amplihack.clone(),
-        BinaryProbe {
-            version: Some("0.9.71".into()),
-            writable: true,
-        },
-    );
-    probes.insert(
-        user_hooks,
-        BinaryProbe {
-            version: Some("0.9.71".into()),
-            writable: true,
-        },
-    );
+    let probes = probe_candidates_without_exec(&report);
 
     let decision = InstallTargetDecision::from(TargetDecisionInput {
         report,
-        current_version: "0.9.71".into(),
         candidate_probes: probes,
         denied_system_prefixes: vec![temp.path().join("usr/local")],
     });
@@ -110,18 +87,10 @@ fn downloaded_update_plan_returns_manual_repair_before_copying_into_denied_syste
         binary_names: vec!["amplihack".into(), "amplihack-hooks".into()],
     })
     .unwrap();
-    let mut probes = BTreeMap::new();
-    probes.insert(
-        current_exe,
-        BinaryProbe {
-            version: Some("0.9.60".into()),
-            writable: false,
-        },
-    );
+    let probes = probe_candidates_without_exec(&report);
 
     let decision = InstallTargetDecision::from(TargetDecisionInput {
         report,
-        current_version: "0.9.71".into(),
         candidate_probes: probes,
         denied_system_prefixes: vec![temp.path().join("usr/local")],
     });
